pyproject.toml
[tool.poetry]
name = "logisco-gitlab-mirror"
version = "1.0.0"
description = "LOGISCO GitLab Mirror - Outil interne de synchronisation GitLab"
authors = ["Darlin TEUMA <hdteuma@logisco.com>"]
readme = "README.md"
packages = [{include = "gitlab_mirror", from = "src"}]
license = "Proprietary"
# repository = "https://votre-gitlab.example.com/groupe/logisco-gitlab-mirror"
keywords = ["gitlab", "sync", "mirror", "logisco"]
classifiers = [
    "Private :: Do Not Upload",
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Version Control :: Git",
]

[tool.poetry.dependencies]
python = "^3.12"
click = "^8.1.7"
python-gitlab = "^4.4.0"
pydantic = "^2.6.0"
pydantic-settings = "^2.1.0"
rich = "^13.7.0"
gitpython = "^3.1.41"
tomli = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
black = "^24.1.0"
ruff = "^0.2.0"
mypy = "^1.8.0"
pre-commit = "^3.6.0"
types-requests = "^2.31.0"

[tool.poetry.scripts]
lgm = "gitlab_mirror.cli:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py312"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = ["E501"]  # line too long (handled by black)

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_calls = true
strict = true

[[tool.mypy.overrides]]
module = "gitlab.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src/gitlab_mirror --cov-report=term-missing"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

